requirements.txt
colour==0.1.5
Pillow==8.3.2
PyQt5==5.15.4
PyQt5-sip==12.9.0
